+++ python/fate_test/fate_test/scripts/testsuite_cli.py
@@ -42,10 +42,6 @@ from fate_test.scripts._utils import _load_testsuites, _upload_data, _delete_dat
               help="When guest, host and flow are deployed on the same machine, the parameter 0 is more appropriate")
 @click.option('-time', '--timeout', type=int, default=3600,
               help="Task timeout duration")
-@click.option('-jp', '--update_job_parameters', default="{}", type=JSON_STRING,
-              help="a json string represents mapping for replacing fields in conf.job_parameters")
-@click.option('-cp', '--update_component_parameters', default="{}", type=JSON_STRING,
-              help="a json string represents mapping for replacing fields in conf.component_parameters")
 @click.option('-iter', '--max_iter', type=int, default=100,
               help="When the algorithm model is LR, the number of iterations is set")
 @click.option('-depth', '--max_depth', type=int, default=4,
@@ -54,6 +50,10 @@ from fate_test.scripts._utils import _load_testsuites, _upload_data, _delete_dat
               help="When the algorithm model is SecureBoost, set the number of trees")
 @click.option('-node', '--processors_per_node', type=int, default=4,
               help="processors per node")
+@click.option('--update_job_parameters', default="{}", type=JSON_STRING,
+              help="a json string represents mapping for replacing fields in conf.job_parameters")
+@click.option('--update_component_parameters', default="{}", type=JSON_STRING,
+              help="a json string represents mapping for replacing fields in conf.component_parameters")
 @click.option("--skip-dsl-jobs", is_flag=True, default=False,
               help="skip dsl jobs defined in testsuite")
 @click.option("--skip-pipeline-jobs", is_flag=True, default=False,
@@ -62,17 +62,21 @@ from fate_test.scripts._utils import _load_testsuites, _upload_data, _delete_dat
               help="skip uploading data specified in testsuite")
 @click.option("--data-only", is_flag=True, default=False,
               help="upload data only")
+@click.option("--disable-clean-data", "clean_data", flag_value=False, default=None)
+@click.option("--enable-clean-data", "clean_data", flag_value=True, default=None)
 @SharedOptions.get_shared_options(hidden=True)
 @click.pass_context
 def run_suite(ctx, replace, include, exclude, glob, timeout, update_job_parameters, update_component_parameters,
-              skip_dsl_jobs, skip_pipeline_jobs, skip_data, data_only, use_local_data, max_iter, max_depth, num_trees,
-              processors_per_node,  **kwargs):
+              skip_dsl_jobs, skip_pipeline_jobs, skip_data, data_only, clean_data, use_local_data, max_iter, max_depth,
+              num_trees, processors_per_node, **kwargs):
     """
     process testsuite
     """
     ctx.obj.update(**kwargs)
     ctx.obj.post_process()
     config_inst = ctx.obj["config"]
+    if clean_data is None:
+        clean_data = config_inst.clean_data
     namespace = ctx.obj["namespace"]
     yes = ctx.obj["yes"]
     data_namespace_mangling = ctx.obj["namespace_mangling"]
@@ -110,7 +114,8 @@ def run_suite(ctx, replace, include, exclude, glob, timeout, update_job_paramete
                     echo.stdout_newline()
                     try:
                         _submit_job(client, suite, namespace, config_inst, timeout, update_job_parameters,
-                                    update_component_parameters, max_iter, max_depth, num_trees, processors_per_node)
+                                    update_component_parameters, max_iter, max_depth, num_trees,
+                                    processors_per_node)
                     except Exception as e:
                         raise RuntimeError(f"exception occur while submit job for {suite.path}") from e
 
@@ -120,7 +125,7 @@ def run_suite(ctx, replace, include, exclude, glob, timeout, update_job_paramete
                     except Exception as e:
                         raise RuntimeError(f"exception occur while running pipeline jobs for {suite.path}") from e
 
-                if not skip_data:
+                if not skip_data and clean_data:
                     _delete_data(client, suite)
                 echo.echo(f"[{i + 1}/{len(suites)}]elapse {timedelta(seconds=int(time.time() - start))}", fg='red')
                 if not skip_dsl_jobs or not skip_pipeline_jobs:
@@ -206,7 +211,22 @@ def _submit_job(clients: Clients, suite: Testsuite, namespace: str, config: Conf
                 job_progress.final(response.status)
                 suite.update_status(job_name=job.job_name, status=response.status.status)
                 if response.status.is_success():
-                    suite.feed_success_model_info(job.job_name, response.model_info)
+                    if suite.model_in_dep(job.job_name):
+                        dependent_jobs = suite.get_dependent_jobs(job.job_name)
+                        for predict_job in dependent_jobs:
+                            if predict_job.job_conf.dsl_version == 2:
+                                # noinspection PyBroadException
+                                try:
+                                    model_info = clients["guest_0"].deploy_model(
+                                        model_id=response.model_info["model_id"],
+                                        model_version=response.model_info["model_version"],
+                                        dsl=predict_job.job_dsl.as_dict())
+                                except Exception:
+                                    _raise()
+                            else:
+                                model_info = response.model_info
+                            suite.feed_dep_model_info(predict_job, job.job_name, model_info)
+                        suite.remove_dependency(job.job_name)
             update_bar(0)
             echo.stdout_newline()
 
@@ -217,27 +237,29 @@ def _run_pipeline_jobs(config: Config, suite: Testsuite, namespace: str, data_na
     for i, pipeline_job in enumerate(suite.pipeline_jobs):
         echo.echo(f"Running [{i + 1}/{job_n}] job: {pipeline_job.job_name}")
 
-        def _raise():
+        def _raise(err_msg, status="failed"):
             exception_id = str(uuid.uuid1())
-            suite.update_status(job_name=job_name, exception_id=exception_id)
-            echo.file(f"exception({exception_id})")
-            LOGGER.exception(f"exception id: {exception_id}")
+            suite.update_status(job_name=job_name, exception_id=exception_id, status=status)
+            echo.file(f"exception({exception_id}), error message:\n{err_msg}")
+            # LOGGER.exception(f"exception id: {exception_id}")
 
         job_name, script_path = pipeline_job.job_name, pipeline_job.script_path
         mod = _load_module_from_script(script_path)
         try:
             if data_namespace_mangling:
                 try:
-                    mod.main(config, f"_{namespace}")
+                    mod.main(config=config, namespace=f"_{namespace}")
                     suite.update_status(job_name=job_name, status="success")
-                except Exception:
-                    pass
+                except Exception as e:
+                    _raise(e)
+                    continue
             else:
                 try:
-                    mod.main(config)
+                    mod.main(config=config)
                     suite.update_status(job_name=job_name, status="success")
-                except Exception:
-                    pass
-        except Exception:
-            _raise()
+                except Exception as e:
+                    _raise(e)
+                    continue
+        except Exception as e:
+            _raise(e, status="not submitted")
             continue
